It works! The graph currently shows all the employees, and colors can be toggled and the order changed to show positions, projects, and teams instead.
<VariableAreaStack/> is the main graph component,
which renders a stacked area graph using a variable number of areas,
as opposed to <FixedAreaStack/>, which uses a fixed number of areas.

Reimplemented the faux dom for performance optimization,
which I made as lightweight as possible (elided many features not used by d3).

Implemented xlsxAsync.readFile() correctly,
which reads the file into a buffer asynchronously,
and then calls xlsx.read() synchronously (as it should).

Added roadmap.txt for future features to add.