Implemented <FilterControls> with for a single item at a time.
Added multiEvent() for detecting double, triple, etc. clicks.

Wrote <VariableAreaStack> as curried functions,
first applying the data, then the size, order, scale, etc.,
and then finally rendering with the color.

The curried nature of <VariableAreaStack> is used in CachedGraph,
which only reloads the necessary parts of <VariableAreaStack>
(i.e. only the final render is called when the color changes,
instead of recalculating the data and size).