Added Re-Scale button:     Normally when you change filter, the graph scale stays the same     so things are easier to visually compare,     but if you click Re-Scale, it scales the new graph to fit the entire graph.
Added CachedGraph, which caches as much of the graph as it can,
only re-rendering what's necessary.
So if only the order is changed, a lot of the calculations can be saved,
and it the coloring is changed, the graph only needs to be recolored in CSS.
This uses a new VariableAreaStack, which is now curried,
so different args can be supplied through successive functions calls,
and intermediate values can be saved instead of recomputed every time.

Added brotli compression at server start time (much better compression that gzip).

Heavily refactored util/ into more directories.

Data is watched for changes, so when the Excel files are updated,
so is the data and rendered app.

Added a blank (for now) login page.

Added SsrRendererFactory,
which creates SsrRendererClasses,
which create SsrRenderers,
which can be have a refresh() method and handler() method to attach to the server.
This is a major refactoring of what used to be ssr.tsx.
Now it's easy to repeat it with the new login/ route.

Added ClientLoader, which hydrates the React component in the browser.
It is separate from SsrRenderer to remove any heavy dependencies,
but SsrRenderer.new() takes a ClientLoader to make it more DRY.

Heavily refactored webpack.config.js.
Now it just imports the exported configs and exports it.
The rest is written in TypeScript in ProjectScheduleViewer/webpack.
This allows webpack.config.js to examine the same SsrRenderers
used in the server and get the htmlPlugins and names from them, etc.

It also exports two configs, a separate client and server one.
I was hoping the server one could run on its own in node,
but it still has bugs, even though it's compiled by webpack for node, not web.